pyproject.toml
[project]
name = "fantasy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.13"
dependencies = [
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.9",
    "aiosignal==1.3.1",
    "altair==5.1.1",
    "async-timeout==5.0.1",
    "attrs==23.1.0",
    "blinker==1.6.2",
    "cachetools==5.3.1",
    "certifi==2023.7.22",
    "charset-normalizer==3.2.0",
    "click==8.1.7",
    "decorator==5.1.1",
    "espn-api==0.43.0",
    "frozenlist==1.5.0",
    "fsspec==2024.10.0",
    "gcsfs==2024.10.0",
    "gitdb==4.0.10",
    "gitpython==3.1.35",
    "google-api-core==2.10.0",
    "google-auth==2.11.0",
    "google-auth-oauthlib==0.5.2",
    "google-cloud==0.34.0",
    "google-cloud-core==2.3.2",
    "google-cloud-storage==2.5.0",
    "google-crc32c==1.5.0",
    "google-resumable-media==2.3.3",
    "googleapis-common-protos==1.56.4",
    "idna==3.4",
    "importlib-metadata==6.8.0",
    "jinja2==3.1.2",
    "jsonschema==4.19.0",
    "jsonschema-specifications==2023.7.1",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.3",
    "mdurl==0.1.2",
    "multidict==6.1.0",
    "numpy==1.25.2",
    "oauthlib==3.2.2",
    "packaging==23.1",
    "pandas==2.1.0",
    "pillow==9.5.0",
    "propcache==0.2.1",
    "protobuf==4.24.3",
    "pyarrow==13.0.0",
    "pyasn1==0.5.0",
    "pyasn1-modules==0.3.0",
    "pydeck==0.8.0",
    "pygments==2.16.1",
    "pympler==1.0.1",
    "python-dateutil==2.8.2",
    "pytz==2023.3.post1",
    "pytz-deprecation-shim==0.1.0.post0",
    "referencing==0.30.2",
    "requests==2.31.0",
    "requests-oauthlib==1.3.1",
    "rich==13.5.2",
    "rpds-py==0.10.2",
    "rsa==4.9",
    "six==1.16.0",
    "smmap==5.0.0",
    "streamlit>=1.49.1",
    "tenacity==8.2.3",
    "toml==0.10.2",
    "toolz==0.12.0",
    "tornado==6.3.3",
    "typing-extensions==4.7.1",
    "tzdata==2023.3",
    "tzlocal==4.3.1",
    "urllib3==2.0.4",
    "validators==0.22.0",
    "yarl==1.18.3",
    "zipp==3.16.2",
]
